Cleanup, copyright and compatibility fixes

+++ performance_test_factory/src/factory.cpp
@@ -1,11 +1,30 @@
-/* Software License Agreement (BSD License)
- *
- *  Copyright (c) 2019, iRobot ROS
- *  All rights reserved.
- *
- *  This file is part of ros2-performance, which is released under BSD-3-Clause.
- *  You may use, distribute and modify this code under the BSD-3-Clause license.
- */
+// Copyright 2019 iRobot ROS
+//
+// Redistribution and use in source and binary forms, with or without
+// modification, are permitted provided that the following conditions are met:
+//
+//    * Redistributions of source code must retain the above copyright
+//      notice, this list of conditions and the following disclaimer.
+//
+//    * Redistributions in binary form must reproduce the above copyright
+//      notice, this list of conditions and the following disclaimer in the
+//      documentation and/or other materials provided with the distribution.
+//
+//    * Neither the name of the iRobot ROS nor the names of its
+//      contributors may be used to endorse or promote products derived from
+//      this software without specific prior written permission.
+//
+// THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS "AS IS"
+// AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE
+// IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE
+// ARE DISCLAIMED. IN NO EVENT SHALL THE COPYRIGHT HOLDER OR CONTRIBUTORS BE
+// LIABLE FOR ANY DIRECT, INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR
+// CONSEQUENTIAL DAMAGES (INCLUDING, BUT NOT LIMITED TO, PROCUREMENT OF
+// SUBSTITUTE GOODS OR SERVICES; LOSS OF USE, DATA, OR PROFITS; OR BUSINESS
+// INTERRUPTION) HOWEVER CAUSED AND ON ANY THEORY OF LIABILITY, WHETHER IN
+// CONTRACT, STRICT LIABILITY, OR TORT (INCLUDING NEGLIGENCE OR OTHERWISE)
+// ARISING IN ANY WAY OUT OF THE USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE
+// POSSIBILITY OF SUCH DAMAGE.
 
 #include <algorithm>
 #include <map>
@@ -549,7 +568,7 @@ void TemplateFactory::add_periodic_publisher_from_json(
     std::cout << "Error! Publishers must set period_ms or freq_hz in json file" << std::endl;
   }
 
-  auto period = std::chrono::microseconds(static_cast<int>(period_ms * 1000));
+  auto period = std::chrono::microseconds(static_cast<int64_t>(period_ms * 1000));
 
   size_t msg_size = 0;
   if (pub_json.find("msg_size") != pub_json.end()) {
@@ -612,7 +631,7 @@ void TemplateFactory::add_periodic_client_from_json(
     std::cout << "Error! Clients must set period_ms or freq_hz in json file" << std::endl;
   }
 
-  auto period = std::chrono::microseconds(static_cast<int>(period_ms * 1000));
+  auto period = std::chrono::microseconds(static_cast<int64_t>(period_ms * 1000));
 
   rclcpp::QoS custom_qos_profile = get_qos_from_json(client_json);
 
@@ -641,7 +660,7 @@ void TemplateFactory::add_periodic_action_client_from_json(
     assert(0 && "Error! Action Clients must set period_ms or freq_hz in json file");
   }
 
-  auto period = std::chrono::microseconds(static_cast<int>(period_ms * 1000));
+  auto period = std::chrono::microseconds(static_cast<int64_t>(period_ms * 1000));
 
   rclcpp::QoS custom_qos_profile = get_qos_from_json(action_client_json);
